Move Models.Th075.Definitions.*Table to the core library
